📝 feat(grade-model): add grade response models
This commit adds two new models: `GradeResModel` and `GradesResModel` to
represent the grade response data from the server. The `GradeResModel`
class defines the structure of a single grade response, while the
`GradesResModel` class contains a list of `GradeResModel` instances.

These models will be used throughout the application to handle grade-
related data and facilitate efficient communication with the backend
server.