✨ feat(routes_manager.dart): Add student exam waiting screen
This commit adds a new route for the student exam waiting screen. The
new route is called `studentExamScreenWaiting` and it is added to the
`Routes` class. The screen is implemented using the `StudentExamWaitingScreen`
class and is bound to the `HomeBindings` and `WebSocketBinding` classes.
This change allows the application to display the student exam waiting
screen when the user navigates to the corresponding route.