✨ feat(home_screen): Add navigation to exam mission screen
Adds a new `InkWell` widget to the `HomeScreen` that navigates the user to the
`StudentExamScreenWaiting` screen when tapped. This allows the user to view
the details of the exam mission they have selected.

The changes focus on the important functionality of allowing the user to
navigate to the exam mission screen from the home screen.