✨ feat(student_seat): add studentSeatNumbers property to StudentsSeatsNumbersResModel
Adds a new `studentSeatNumbers` property to the `StudentsSeatsNumbersResModel`
class. This property will hold a list of `StudentSeatNumberResModel` objects,
which represent the available student seat numbers.

The changes were made to ensure that the `StudentsSeatsNumbersResModel` class
correctly represents the data being returned from the API, which includes the
list of student seat numbers.