📦️ feat: Update dependencies and add package_info_plus
This commit updates the dependencies in the `pubspec.lock` file and adds the `package_info_plus` package.

The key changes are:
- Update the `vm_service` package from version `14.2.4` to `14.2.5`.
- Add the `win32` package with version `5.5.4`.
- Add the `package_info_plus` package with version `8.0.2`.
- Add the `package_info_plus_platform_interface` package with version `3.0.1`.

These changes are necessary to ensure the project is using the latest stable versions of the dependencies and to add the `package_info_plus` package, which will provide information about the application package, such as the version and build number.