🔥 feat(links): update production and staging URLs
Updates the production and staging URLs in the `app_links.dart` file to match the latest deployments.

The production URL has been updated to `https://control-2025-490607372730.us-central1.run.app/`, and the staging URL has been updated to `https://control-2025-904810505533.us-central1.run.app/`.

This change ensures that the application is pointing to the correct backend endpoints for both production and staging environments.