```
feat(home-controller): add server clock model

The changes in this commit are focused on adding a new `ServerClockModel` to the home controller. This model is used to handle the server clock, which is important for accurately displaying exam times and other time-related information.

The key changes are:

1. Import the `ServerClockModel` in the home controller.
2. Add a new property `serverClockModel` to the home controller.
3. Initialize the `serverClockModel` in the home controller's constructor.

These changes will enable the home controller to better manage the server clock and provide accurate time-related information to the users.
```