Remove widget_test.dart file
The `widget_test.dart` file, which contained a basic Flutter widget test, has been deleted. This test was used to verify the functionality of the counter widget in the application. The test ensured that the counter started at 0, incremented correctly when the '+' icon was tapped, and displayed the updated value. The removal of this file suggests that the test is no longer needed or has been replaced by a more comprehensive testing strategy.